Calculate hash differently
The default `cache-key` value was static if the `committish` input
didn't change, which could result in stale caches. The new default value
is now derived from both the `github.sha` context value and the
`committish` input value.

This guarantees that every new commit that triggers the CI will always
store its files to be used in the next CI run.

GitHub Actions cache will just prune old cache entries if they are too
old or there is not enough space.